refactor pipette files
