+++ shared-data/pipette/schemas/model/1.json
@@ -79,7 +79,7 @@
           "additionalProperties": false,
           "properties": {
             "name": {
-              "description": "reference to name of this version, should match a key in pipetteNameSpecs.json",
+              "description": "reference to name of this version, should match a key in corresponding pipette 'name' json file",
               "type": "string"
             },
             "top": { "$ref": "#/definitions/editConfigurations" },
